core: utils mod with SliceReader and BufWriter
Refactor the `SliceReader` into `core::utils`, so that it can be re-used.
